Change settings for RegisterPage.js in SetRegisterPage.js.js and move some functions and objects
