+++ Front/clinicfront/src/components/RegisterPage.js
@@ -1,7 +1,11 @@
 import React from "react";
 import {Button, Col, Form} from "react-bootstrap";
 import {useFormFields} from "../containers/SetRegisterPage"
-import {sendRequestByGivenDetails} from "../actions";
+
+import {
+    sendFetchRequest,
+    styleForMainDiv, styleForForm, styleForFormLabel
+} from "../containers/SetRegisterPage";
 
 
 export const RegisterPage = () => {
@@ -19,69 +23,13 @@ export const RegisterPage = () => {
         setUserRegisterInformation(event);
     };
 
-    //CSS stylesheet
-    const styleForFormLabel = {color:'white'};
-
-    const styleForForm = {
-        border: '2px solid white',
-        borderRadius: '5px',
-        padding: '8px'
-    };
-
-    const styleForMainDiv = {
-        margin: '50px auto auto 50px',
-        width: '60%',
-    };
-
-    //Fetch packages
-    const specialFunctionSendRequestByGivenDetails = (responseJSONData) => {
-        const body = (userRegisterInformation) => ({
-            doctoruuid: responseJSONData.uuid,
-            firstName: userRegisterInformation.firstName,
-            lastName: userRegisterInformation.lastName,
-            photoUrl: userRegisterInformation.photoURL,
-            licence: userRegisterInformation.licence
-        });
-
-        const headers = {
-            'Authorization': responseJSONData.token,
-            'Content-Type': 'application/json;charset=UTF-8',
-        };
-
-        sendRequestByGivenDetails(
-            "http://localhost:8762/doctor-mssc/doctors/",
-            'POST',
-            body(userRegisterInformation),
-            headers,
-            null,
-            null
-        );
-    };
-
-    const fetchRequest = () => {
-        const body = (userRegisterInformation) => ({
-            "email": userRegisterInformation.email,
-            "password": userRegisterInformation.password,
-            "role": userRegisterInformation.role
-        });
-
-        sendRequestByGivenDetails(
-            "http://localhost:8762/auth/users/",
-            'POST',
-            body(userRegisterInformation),
-            {'Content-Type': 'application/json'},
-            null,
-            specialFunctionSendRequestByGivenDetails
-        );
-    };
-
     //Main HTML return
     return(
         <div style={styleForMainDiv}>
             <Form
                 onSubmit={e => {
                     e.preventDefault();
-                    fetchRequest();
+                    sendFetchRequest(userRegisterInformation)
                 }}
                 style={styleForForm}
             >
